+++ CHANGES.md
@@ -16,6 +16,7 @@ Changes in 11.3
 - Fix incoherence with AllowEdit() when category_id present or not in URL in Student.php & User.php
 - Refuse to enroll student twice in the same course period in Schedule.php & MassSchedule.php
 - Update French & Spanish translations in rosariosis.po & help.po
+- Add Percent grade inside HTML comment so we can accurately sort by Grade column in InputFinalGrades.php
 
 Changes in 11.2.4
 -----------------